feat: add endpoints splitting
